+++ src/main/java/pl/tradeengine/application/dto/PriceUpdateDto.java
@@ -1,10 +1,15 @@
 package pl.tradeengine.application.dto;
 
+import java.math.BigDecimal;
+
 public record PriceUpdateDto(
         String symbol,
         String timeframe,
         String signalType,
-        double currentHigh,
-        double currentLow,
-        Double close
+//        double currentHigh,
+//        double currentLow,
+//        Double close
+        BigDecimal currentHigh, // <-- ZMIEŃ TYP
+        BigDecimal currentLow,  // <-- ZMIEŃ TYP
+        BigDecimal close
 ) {}
